+++ Aura/Sampler.cs
@@ -63,8 +63,8 @@ namespace Aura
 
     class Sample
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
 
         public float OffsetX { get; set; }
         public float OffsetY { get; set; }
